refactor: AnalyzeManager 수정
