+++ GyroData/Manager/AnalysisManager.swift
@@ -16,22 +16,22 @@ protocol AnalysisManagerType {
 }
 
 struct AnalysisManager: AnalysisManagerType {
-    let manager: AnalysisManagerType
-    
-    init(analysis: AnalysisType) {
-        switch analysis {
-        case .accelerate:
-            manager = AccelerateManager()
-        case .gyroscope:
-            manager = GyroscopeManager()
-        }
-    }
+    private let motion = CMMotionManager()
     
     func startAnalyse() -> AnalysisData {
-        return manager.startAnalyse()
+        guard motion.isAccelerometerAvailable else {
+            return (x: 0, y: 0, z: 0)
+        }
+        motion.accelerometerUpdateInterval = 0.1
+        motion.startAccelerometerUpdates()
+        
+        guard let data = self.motion.accelerometerData?.acceleration else {
+            return (x: 0, y: 0, z: 0)
+        }
+        return (x: data.x, y: data.y, z: data.z)
     }
     
     func stopAnalyse() {
-        return manager.stopAnalyse()
+        motion.stopAccelerometerUpdates()
     }
 }
